+++ AgentBasedModel/__init__.py
@@ -1,5 +1,16 @@
-from AgentBasedModel.agents import *
-from AgentBasedModel.events import *
-from AgentBasedModel.simulator import *
+from AgentBasedModel.extra import *
 from AgentBasedModel.visualization import *
-from AgentBasedModel.states import *
+
+from AgentBasedModel.exchange import (
+    ExchangeAgent
+)
+from AgentBasedModel.traders import (
+    Random,
+    Fundamentalist,
+    Chartist,
+    Universalist
+)
+from AgentBasedModel.simulator import (
+    Simulator,
+    SimulatorInfo
+)
