--- AgentBasedModel/agents/__init__.py
@@ -1,2 +1,0 @@
-from AgentBasedModel.agents.agents import ExchangeAgent, Trader, Random, Chartist, Fundamentalist, Universalist,\
-    MarketMaker
