--- AgentBasedModel/simulator/__init__.py
@@ -1,1 +1,0 @@
-from AgentBasedModel.simulator.simulator import Simulator, SimulatorInfo
