--- AgentBasedModel/states/__init__.py
@@ -1,1 +1,0 @@
-from AgentBasedModel.states.states import aggToShock, general_states
