+++ AgentBasedModel/visualization/__init__.py
@@ -1,4 +1,20 @@
-from AgentBasedModel.visualization.market import plot_price, plot_price_fundamental, plot_arbitrage, plot_dividend,\
-    plot_orders, plot_volatility_price, plot_volatility_return, plot_liquidity
-from AgentBasedModel.visualization.trader import plot_equity, plot_cash, plot_assets, plot_returns,\
-    plot_strategies, plot_strategies2, plot_sentiments, plot_sentiments2
+from AgentBasedModel.visualization.market import (
+    plot_price,
+    plot_price_fundamental,
+    plot_arbitrage,
+    plot_dividend,
+    plot_orders,
+    plot_volatility_price,
+    plot_volatility_return,
+    plot_liquidity
+)
+from AgentBasedModel.visualization.trader import (
+    plot_equity,
+    plot_cash,
+    plot_assets,
+    plot_returns,
+    plot_strategies,
+    plot_strategies2,
+    plot_sentiments,
+    plot_sentiments2
+)
