+++ main.py
@@ -1,10 +1,13 @@
 from AgentBasedModel import *
-from AgentBasedModel.utils.math import *
 
 exchange = ExchangeAgent(volume=1000)
 simulator = Simulator(**{
     'exchange': exchange,
-    'traders': [Fundamentalist(exchange, 10**3) for _ in range(10)],
+    'traders': [
+        *[Random(exchange, 10**3)         for _ in range(10)],
+        *[Fundamentalist(exchange, 10**3) for _ in range(10)],
+        *[Chartist(exchange, 10**3)       for _ in range(10)]
+    ],
     'events': [MarketPriceShock(200, -20)]
 })
